HTML tutorial/Backend2/requirements.txt
Flask
Pillow
python-dotenv
google-generativeai